wip: move interchain test helpers into separate file
